Preparing to switch browse-worker to another user.
# Problem
It is really annoying that I have to do a clean build whenever I want to build & run the NewtonCommander app.

# Long explanation why this is a problem
Currently I SETUID the browse-worker from the NewtonCommander app.
With Cocoapods I can no longer do that.

Cocoapods has deprecated pre_install and post_install, so I cannot compile the browse-worker in this step.
Instead I have to place a compiled version of the browse-worker in a NewtonCommanderBrowse.bundle
When I compile the entire NewtonCommander app, then this NewtonCommanderBrowse.bundle is copied into the Resources folder.
When running the NewtonCommander app then I change permissions of the browse-worker within the NewtonCommanderBrowse.bundle.
All this works from a clean build.

However if I have not cleaned then the NewtonCommanderBrowse.bundle cannot be overwritten because of its altered permissions.

# Solution
Instead have a command line util that can start the NewtonCommanderBrowse.bundle/browse-worker with the right permissions.